+++ BubblesDivePlanner/Controllers/IDiveController.cs
@@ -6,7 +6,7 @@ namespace BubblesDivePlanner.Controllers
     public interface IDiveController
     {
         void SetupDivePlan(IDivePlan divePlan);
-        IDivePlan SetupDiveStep(byte depthCeiling);
+        IDivePlan SetupDiveStep(double depthCeiling);
         IDivePlan RunDiveProfile(IDivePlan divePlan);
         void PrintDiveResult(IDivePlan divePlan);
     }
